% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe graphics}
\arguments{
\item{lhs, rhs}{A visualisation and a function to apply to it}
}
\description{
Like dplyr, ggvis also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
# Instead of
dplyr::mutate(dplyr::filter(iris, Species == "versicolor"),
"Sample" = paste0(Species, dplyr::row_number()))
# You can write
dplyr::filter(iris, Species == "versicolor") \%>\%
dplyr::mutate("Sample" = paste0(Species, "_", dplyr::row_number()))
}
