% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrouper.R
\name{plotGrouper}
\alias{plotGrouper}
\title{A function to run the shiny app}
\usage{
plotGrouper(...)
}
\arguments{
\item{...}{Any argument that you can pass to shiny::runApp}
}
\description{
This function allows you to run the app
}
\examples{
plotGrouper()
}
