% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{A function to read the data to be plotted into the app}
\usage{
readData(sheet = NULL, file = NULL)
}
\arguments{
\item{sheet}{Takes a vector of sheets to be read}

\item{file}{Takes an excel file to be read from}
}
\value{
Tibble assembled from the sheets selected from the file
}
\description{
This function will read each sheet listed and return a tibble
}
\examples{
readData()
}
\keyword{readData}
