% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeData.R
\name{organizeData}
\alias{organizeData}
\title{A function to organize the raw data to be plotted}
\usage{
organizeData(data = NULL, exclude = NULL, comp = NULL,
  comps = NULL, variables = NULL, id = NULL, beadColumn = NULL,
  dilutionColumn = NULL)
}
\arguments{
\item{data}{Takes a tibble}

\item{exclude}{Takes list of columns to exclude from gather}

\item{comp}{Takes name of comparison column}

\item{comps}{Takes a vector of names of the comparisons}

\item{variables}{Takes vector of the variables to be plotted}

\item{id}{Takes name of unique identifier column}

\item{beadColumn}{Takes column name that has total number of beads/sample}

\item{dilutionColumn}{Takes column name that has dilution factor for each sample 1/x}
}
\description{
This function will organize the data and perform count calculations if appropriate
}
\examples{
iris \%>\% mutate(Species = as.character(Species)) \%>\%
group_by(Species) \%>\%
mutate(Sample = paste0(Species, "_", row_number()), Sheet = "iris") \%>\%
select(Sample, Sheet, Species, everything()) \%>\%
organizeData(data = .,
exclude = c("Sample", "Sheet", "Species"),
comp = "Species",
comps = c("setosa", "versicolor", "virginica"),
variables = "Sepal.Length",
id = "Sample",
beadColumn = "none",
dilutionColumn = "none")
}
\keyword{organizeData}
