% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeData.R
\name{organizeData}
\alias{organizeData}
\title{A function to organize the raw data to be plotted}
\usage{
organizeData(data, exclude, comp, comps, variables, id, bead, dilution,
  count)
}
\arguments{
\item{data}{Takes a tibble}

\item{exclude}{Takes list of columns to exclude from gather}

\item{comp}{Takes name of comparison column}

\item{comps}{Takes a vector of names of the comparisons}

\item{variables}{Takes vector of the variables to be plotted}

\item{id}{Takes name of unique identifier column}

\item{bead}{Takes number of beads per sample}

\item{dilution}{Takes dilution factor}

\item{count}{True or False indicating presence of columns specifying Total Bead and Dilution}
}
\description{
This function will organize the data and perform count calculations if appropriate
}
\examples{
organizeData()
}
\keyword{organizeData}
