% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrouper.R
\name{plotGrouper}
\alias{plotGrouper}
\title{A function to run the shiny app}
\usage{
plotGrouper(...)
}
\arguments{
\item{...}{Takes any arguments that can be called with shiny::runApp()}
}
\description{
This function allows you to run the app
}
\examples{
plotGrouper()
}
\keyword{plotGrouper}
